refactor(dotnet-sdk): move controllers to V3
Move all controllers into `Controllers/V3/` folder/namespace.

Related-Task: INTER-1878
