+++ projects/dotnet-sdk/Controllers/V3/VisitsController.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Mvc;
 using FingerprintPro.ServerSdk.Model;
 using dotnet_sdk.Models;
 
-namespace dotnet_sdk.Controllers;
+namespace dotnet_sdk.Controllers.V3;
 
 [ApiController]
 [Route("")]
